renaming model files
