+++ src/models/index.ts
@@ -1,5 +1,5 @@
-export * from './models.globsorganizer';
-export * from './models.options';
-export * from './models.optionsmetadata';
-export * from './models.packagedefinition';
-export * from './models.state';
+export * from './GlobsOrganizer';
+export * from './Options';
+export * from './OptionsmMetadata';
+export * from './PackageDefinition';
+export * from './State';
